+++ test-optimizable.js
@@ -1,12 +1,15 @@
-var assert  = require('assert')
-var optimal = require('@optimal/fn')
-var tasks   = require('./index.js')
+'use strict'
+
+const assert  = require('assert')
+const optimal = require('@optimal/fn')
+const tasks   = require('./index.js')
 
 describe('verify optimizability', function() {
 
   it('with basic args', function(done) {
-    var count = 3
-    var result = optimal(tasks, null, [
+    let count = 3
+
+    const result = optimal(tasks, null, [
       {},
       [
         function(next) { next() }